Add files via upload
Committing all project files as a starting point. I have implemented the ofApp::setup() and ofApp::draw() methods - they seem to be working. (Particles appear in window, as do obstacles).

 I started working on RoundEntity::isCollidingWithRoundEntity(const RoundEntity& entity)  ->  I used the geometric distance formula to compute whether the distance between two entities is less than the sum of their radii.

I started the RoundParticle::doAttractToMouse() method. This one is not working yet.

I started both of the RoundParticle::doCollisionWithRoundEntity() methods. As of right now, whenever a particle collides with another entity, it's x and y velocity components both simply reverse. <4, 8> would become <-4, -8>. This doesn't seem to be the behavior exhibited in the demo video. These particles (in the video) seem to be bouncing off in more randomized directions and speeds. This part needs some work.