Add files via upload
Finally got the collision response functions working the way I wanted. I get the vector normal to the collision, record the "speed" of the particle at time of collision. The normal vector is added to the velocity vector, which is then normalized. Finally, the new velocity vector is multiplied by the original speed. Thats it!